+++ Plugin/CodeRebirth/src/Patches/RoundManagerPatch.cs
@@ -568,12 +568,12 @@ static class RoundManagerPatch {
 	[HarmonyPatch(nameof(RoundManager.UnloadSceneObjectsEarly)), HarmonyPostfix]
 	private static void PatchFix_DespawnOldCrates()
 	{
-		foreach (ItemCrate crate in GameObject.FindObjectsOfType<ItemCrate>())
+		foreach (ItemCrate crate in ItemCrate.Instances)
 		{
 			crate.NetworkObject.Despawn();
 		}
 
-		foreach (BiomeManager biome in GameObject.FindObjectsOfType<BiomeManager>())
+		foreach (BiomeManager biome in BiomeManager.Instances)
 		{
 			biome.NetworkObject.Despawn();
 		}